+++ app/src/main/java/io/github/stcksmsh/kap/data/WaterIntake.kt
@@ -7,7 +7,7 @@ import java.util.Date
 @Entity(tableName = "water_intake")
 data class WaterIntake(
     @PrimaryKey(autoGenerate = true) val id: Int = 0,
-    val intakeAmount: Float, // Amount of water in milliliters
+    val intakeAmount: Double, // Amount of water in milliliters
     val date: Date = Date(), // Date of the intake
 ) {
     override fun toString(): String {
