+++ app/src/main/java/io/github/stcksmsh/kap/data/WaterIntakeDao.kt
@@ -24,12 +24,12 @@ interface WaterIntakeDao {
     @Query(
         "SELECT SUM(intakeAmount) FROM water_intake WHERE DATE(date / 1000, 'unixepoch') = DATE(:date / 1000, 'unixepoch')"
     )
-    fun getTodaysIntake(date: Date = Date()): Flow<Float>
+    fun getTodaysIntake(date: Date = Date()): Flow<Double>
 
     @Query(
         "SELECT SUM(intakeAmount) FROM water_intake WHERE DATE(date / 1000, 'unixepoch') = DATE(:date / 1000, 'unixepoch')"
     )
-    fun getTodaysIntakeValue(date: Date = Date()): Float
+    fun getTodaysIntakeValue(date: Date = Date()): Double
 
     @Query("SELECT * FROM water_intake ORDER BY date DESC LIMIT 1")
     fun getLastIntake(): WaterIntake?
