+++ app/src/main/java/io/github/stcksmsh/kap/model/UserSettings.kt
@@ -2,7 +2,7 @@ package io.github.stcksmsh.kap.model
 
 data class UserSettings(
     val age: Int,
-    val weight: Float,
+    val weight: Double,
     val dailyPhysicalActivity: Int,
-    val dailyWaterGoal: Float,
+    val dailyWaterGoal: Double,
 )
